Fixed a few small bugs and added power up system
If a player's rebounded beam makes contact with 3 other enemy beams, the
enemy beams will be destroyed and the players beam will be upgraded more
powerful version of the beam that will have a much higher chance of
piercing the boss' shield. This is to encourage control of the beam when
the player rebounds it.